conf: separate out migration
