+++ conf/mksyscall.go
@@ -5,4 +5,4 @@
 
 package conf
 
-//go:generate go run $GOROOT/src/syscall/mksyscall_windows.go -output zsyscall_windows.go dnsresolver_windows.go path_windows.go storewatcher_windows.go
+//go:generate go run $GOROOT/src/syscall/mksyscall_windows.go -output zsyscall_windows.go dnsresolver_windows.go migration_windows.go storewatcher_windows.go
